pyproject.toml
[project]
name = "llm-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markitdown[docx,pdf]>=0.1.2",
    "openai>=1.99.6",
    "pymilvus>=2.6.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
